More actor float

+++ src/Engine/Objects/Actor.h
@@ -133,7 +133,7 @@ class Actor {
     static void AI_SpellAttack(unsigned int uActorID, struct AIDirection *pDir,
                                SpellId uSpellID, ActorAbility a4, CombinedSkillValue uSkill);
     static void ActorDamageFromMonster(Pid attacker_id, unsigned int actor_id,
-                                       Vec3i *pVelocity, ActorAbility a4);
+                                       Vec3f *pVelocity, ActorAbility a4);
 
     static unsigned short GetObjDescId(SpellId spellId);
 
@@ -155,9 +155,8 @@ class Actor {
     int _43B3E0_CalcDamage(ActorAbility dmgSource);
     static void AddOnDamageOverlay(unsigned int uActorID, int overlayType, int damage);
 
-    static void Arena_summon_actor(MonsterId monster_id, Vec3i pos);
-    static void DamageMonsterFromParty(Pid a1, unsigned int uActorID_Monster,
-                                       Vec3i *pVelocity);
+    static void Arena_summon_actor(MonsterId monster_id, Vec3f pos);
+    static void DamageMonsterFromParty(Pid a1, unsigned int uActorID_Monster, Vec3f *pVelocity);
     static void MakeActorAIList_ODM();
     static int MakeActorAIList_BLV();
     static void UpdateActorAI();
@@ -216,8 +215,8 @@ class Actor {
     uint16_t pitchAngle = 0;
     int sectorId = 0;
     Duration currentActionLength = 0_ticks;
-    Vec3i initialPosition;
-    Vec3i guardingPosition;
+    Vec3f initialPosition;
+    Vec3f guardingPosition;
     uint16_t tetherDistance = 256;
     AIState aiState = Standing;
     ActorAnimation currentActionAnimation = ANIM_Standing;
@@ -262,7 +261,7 @@ void SpawnEncounter(struct MapInfo *pMapInfo, SpawnPoint *spawn, int a3, int a4,
 void evaluateAoeDamage();
 double sub_43AE12(signed int a1);
 void ItemDamageFromActor(Pid uObjID, unsigned int uActorID,
-                         Vec3i *pVelocity);
+                         Vec3f *pVelocity);
 
 // TODO: in original binary almost all calls are with appendOnly=true, only Spawn_Light_Elemental uses
 // appendOnly=false. And this actually makes sense as actor ids can be stored in all kinds of places (e.g. inside
