+++ src/GUI/UI/NPCTopics.cpp
@@ -361,7 +361,7 @@ void prepareArenaFight(DialogueId dialogue) {
 
     for (int i = 0; i < monstersNum; ++i) {
         Vec2i pos = pMonsterArenaPlacements[i];
-        Actor::Arena_summon_actor(grng->randomSample(monsterIds), Vec3i(pos.x, pos.y, 1));
+        Actor::Arena_summon_actor(grng->randomSample(monsterIds), Vec3f(pos.x, pos.y, 1));
     }
     pAudioPlayer->playUISound(SOUND_51heroism03);
 }
